obs-browser: remove force fps sync to global fps
Previously, FPS was forced to synchronize to OBS global FPS for newer
versions of CEF to overcome a rendering bug which was discovered to be
associated with WasHidden(true) calls.

Since WasHidden(true) / WasHidden(false) calls were replaced with page
visibility API simulation for those versions of CEF exhibiting the bug,
this workaround is no longer necessary.